@@ -3,7 +3,7 @@ const Conventions = {
     Deepfake: [
       [
         'Государства — участники настоящей Конвенции',
-        'с целью стимуляции созидательной культуры в противовес потребительской',
+        '',
         'с целью поддержания вовлеченности в настоящую действительность',
         'с целью поддержания психического здоровья современного человека и защиты психического здоровья ребенка',
         'считая, что за неимением исчерпывающих исследований на тему влияния гиперреальной виртуальной реальности на сознание, психическое состояние и здоровье человека, и считая, что в таком случае имеется необходимость во временном ограничении развития данной технологии',
@@ -82,4 +82,87 @@ const Conventions = {
   English: {},
 };
 
-module.exports = { data: Conventions };
+export default Conventions;
+
+
+/* 
+const deepfake = {
+  Preamble: [
+    '',
+    '',
+    '',
+    '',
+    '',
+    '',
+    '',
+  ],
+  I: {
+    "Article-1": {
+      A1: [
+        '',
+      ],
+      A2: [
+        '',
+      ],
+    },
+  },
+  II: {
+    "Article-2": {
+      A3: [
+        '',
+        '',
+      ],
+      A4: [
+        '',
+        '',
+        '',
+      ],
+      A5: [
+        '',
+        '',
+      ],
+      A6: [
+        '',
+        '',
+      ],
+    },
+  },
+  III: {
+    "Article-3": {
+      A7: [
+        '',
+      ],
+      A8: [
+        '',
+      ],
+    },
+  },
+  IV: {
+    "Article-4": {
+      A9: [
+        '',
+      ],
+      A10: [
+        '',
+        '',
+        '',
+        '',
+      ],
+      A11: [
+        '',
+        '',
+      ],
+      A12: [
+        '',
+        '',
+        '',
+      ],
+      A13: [
+        '',
+        '',
+      ],
+    },
+  },
+};
+
+*/
